KrümelBot/requirements.txt
aiohttp
aiohttp
appdirs
argparse
astroid
astroid
async_timeout
async_timeout
atomicwrites
atomicwrites
attr
attrs
certifi
certifi
chardet
chardet
colorama
colorama
daemonize
discord
discord
dotenv
idna
idna
importlib_metadata
importlib_metadata
isort
isort
lazy_object_proxy
lazy_object_proxy
lxml
lxml
mccabe
more_itertools
more_itertools
multidict
multidict
mysql_replication
packaging
packaging
pg_chameleon
pip
pip
pkg_resources
pluggy
pluggy
psycopg2
psycopg2_binary
py
py
pygubu
pygubu
pygubudesigner
pylint
pylint
pymysql
PyMySQL
pymysqlreplication
pyparsing
pytest
python_dotenv
PyYAML
requests
requests
rollbar
rollbar
setuptools
setuptools
six
tabulate
typed_ast
typed_ast
urllib3
urllib3
wcwidth
wcwidth
websockets
websockets
wrapt
wrapt
yaml
yarl
yarl
zipp
